requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==1.16.0
dotenv==0.9.9
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyJWT==2.10.1
pymongo==3.12.0
python-dotenv==1.0.1
Werkzeug==3.1.3
